Add transformers_cache to scripts and .gitignore
Added a line to train_cosine.sh and train.sh to declare and export TRANSFORMERS_CACHE, which is used to specify the location for storing pre-trained transformer models. Such a specification helps to manage, re-use and avoid unnecessary re-downloads of these large models. Also added it to .gitignore so that these cache files are not tracked by Git, preventing unnecessary bloat in repositories.