Add functionality to find and apply optimal threshold for predictions
This commit introduces a new file 'find_threshold.py' which has a function to find the optimal threshold for model predictions using F1 score. This threshold is then applied to the main model's predictions in 'main.py' before training/testing. It also adds a 'predict_step' method in 'sbert_model.py' to make the model work with the Trainer's .predict() method. Additionally, a 'predict_dataloader' function is added in 'datamodule.py'. The compile_model argument in 'sbert_model.py' and the '--compile' option in training scripts ('train.sh') allow users to enable or disable model compilation for SentenceTransformers model.